V31.0.4 (#11)
* Add a chance of a daily games being a mystery (#748)

* Add a 5% chance of the daily game being a mystery.
Add a message to VTRomInfo.vue informing the player that it is a mystery game.  Translated in FR, DE, and ES.
This is setup in a way where we could eventually have spoiler dailies, however it is not turned on at this time.  Generate is weighted at 0 because it won't work at this time and instead just behave like spoilers were on.

* We don't need the $spoilers variable

* Fix Vanilla Pedestal Sword (#747)

* Don't overflow silvers for swordless seeds.

* Fix vanilla pedestal sword.

* Add unit tests for Vanilla Swords.

* Text suggestions from the community. (#746)

* Capitalization

* Uncle, Blind, and Hint additions.

* Recommended changes to uncle's text.

* Silvers hint in Hard+ (#757)

* Make Progressive Bow work with Retro

Progressive Bow broke Retro Logic by overwriting the need for the ShopArrow.
Fixed that.

* Changing Tabs to Spaces

The important stuff

* Add $world to canShootArrows

Give $world variable with every canShootArrows.

$world variable is used to skip shoparrow check in games without rupee bow (aka Retro or customizer with that setting)

* Added UnitTest for Retro Bow working

Basically checking if you can shoot Arrows in Retro where Bow is Hardrequired.

* Cutting out useless stuff

Stuff I had copied and wouldn't need...

* Added Parameters to the Doc Block

Added the new Parameters to the Doc Block

Added an old Parameter that wasn't in before

* Removing Silvers Hint in Hard+

Removes the Silver hint if Progressive Bow gets caped at 1.
Also Comments on Chat not giving Silvers in Crowd Control.

* Silver Arrow GT Text

Adds switch() for Progressive Bows in GT and changes the Text for them to "my Tower."

(Readded Function)

* set overflow for Bow to 2 before the item pool changes

Forgot to add this.
Overflow Count isn't already Pre-Set, so if all cases are skipped, it's essentially 0 and <2 always gives out true.

* Generate Button Text rewording (#758)

* Make Progressive Bow work with Retro

Progressive Bow broke Retro Logic by overwriting the need for the ShopArrow.
Fixed that.

* Changing Tabs to Spaces

The important stuff

* Add $world to canShootArrows

Give $world variable with every canShootArrows.

$world variable is used to skip shoparrow check in games without rupee bow (aka Retro or customizer with that setting)

* Added UnitTest for Retro Bow working

Basically checking if you can shoot Arrows in Retro where Bow is Hardrequired.

* Cutting out useless stuff

Stuff I had copied and wouldn't need...

* Added Parameters to the Doc Block

Added the new Parameters to the Doc Block

Added an old Parameter that wasn't in before

* Generate Button Rewording

Reword "Generate ROM" and "Generate ROM (with spoilers)" to include their usecase.

This was a complaint for some time now and should help reduce the amount of People generating Race ROM's on accident.